+++ spec/evilution/mutator/registry_spec.rb
@@ -4,11 +4,11 @@ RSpec.describe Evilution::Mutator::Registry do
   let(:registry) { described_class.new }
 
   describe ".default" do
-    it "returns a registry pre-loaded with all 18 operators" do
+    it "returns a registry pre-loaded with all 19 operators" do
       default_registry = described_class.default
 
       expect(default_registry).to be_a(described_class)
-      expect(default_registry.operator_count).to eq(18)
+      expect(default_registry.operator_count).to eq(19)
     end
 
     it "includes all expected operator classes" do
@@ -33,7 +33,8 @@ RSpec.describe Evilution::Mutator::Registry do
         Evilution::Mutator::Operator::MethodBodyReplacement,
         Evilution::Mutator::Operator::NegationInsertion,
         Evilution::Mutator::Operator::ReturnValueRemoval,
-        Evilution::Mutator::Operator::CollectionReplacement
+        Evilution::Mutator::Operator::CollectionReplacement,
+        Evilution::Mutator::Operator::MethodCallRemoval
       ]
 
       expect(operators).to match_array(expected_operators)